third_party/blink/web_tests/virtual/web-components-v0-disabled/fast/dom/custom/callback-timing-expected.txt

CONSOLE ERROR: line 22: Uncaught TypeError: document.registerElement is not a function
Ensure that custom element callback fires before DOMContentLoaded event

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

FAIL successfullyParsed should be true. Was false.

TEST COMPLETE

FAIL eventlog[0] should be created x-foo. Was DOMContentLoaded event.
FAIL eventlog[1] should be DOMContentLoaded event. Was load event.
FAIL eventlog[2] should be load event (of type string). Was undefined (of type undefined).
